feat(ai-models): tabbed Overview/Pricing/Benchmarks/Family + e2e coverage
Layered improvements on top of the EntityProfileShell migration:

- Decompose /ai-models/[slug] into Overview/Pricing/Benchmarks/Family tabs
  via the shell's `tabs` prop. Tabs hide automatically when their data is
  empty (a model with no benchmarks won't show the Benchmarks tab, etc.).
  Sidebar now hosts only Details + Tags — Capabilities moves into the
  Overview tab so the sidebar stays focused on canonical metadata.

- New apps/web/src/app/ai-models/[slug]/tabs.tsx with the four tab
  components co-located with the page. page.tsx drops from 459 → 285
  lines as the tab JSX moves to the dedicated module.

- Extend e2e/render-audit.spec.ts to cover the new ai-model detail pages:
  3 model slugs added to TABBED_PAGES (claude-opus-4-5, gemini-2-5-pro,
  gpt-4o) and 2 to STAT_CARD_PAGES so the empty-stat-card check runs on
  the newly tabbed layout.

Verified locally: clicked through all 4 tabs on claude-opus-4-5,
gemini-2-5-pro, deepseek-v3 — each panel shows distinct content
(modality+capabilities, pricing table, benchmark scorecard, family table)
with zero console errors. orgs/anthropic, orgs/anthropic/data, and
people/dario-amodei still render cleanly.

Refs QUA-328
